Update to move detail inspection window
Detail inspection window is now hidden until brought to the
front, it is also only accessible from the trips page.
