+++ internal/gui/caves.go
@@ -38,8 +38,8 @@ func (c *caves) setKeybinding(g *Gui) {
 		g.setGlobalKeybinding(event)
 
 		switch event.Key() {
-		case tcell.KeyEnter:
-			g.inspectCave()
+		//case tcell.KeyEnter:
+		//	g.inspectCave()
 		case tcell.KeyCtrlR:
 			c.setEntries(g)
 		}
