+++ internal/gui/navigate.go
@@ -11,9 +11,10 @@ func newNavigate() *navigate {
 	navi := &navigate{
 		TextView: tview.NewTextView().SetTextColor(tview.Styles.PrimaryTextColor),
 		keybindings: map[string]string{
-			"trips" : " n: New Log Entry, m: Modify Log,  d: Remove Log, /: filter, Enter: Inspect ",
-			"caves" : " n: New Cave, m: Modify Cave, d: Remove Cave, /: filter, Enter: Inspect ",
-			"cavers": " n: New Caver, m: Modify Caver, d: Remove Caver, /: filter, Enter: Inspect ",
+			"trips" : " n: New Log Entry, m: Modify Log,  d: Remove Log, /: Filter, Enter: Inspect Detail ",
+			"caves" : " n: New Cave, m: Modify Cave, d: Remove Cave, /: Filter ",
+			"cavers": " n: New Caver, m: Modify Caver, d: Remove Caver, /: Filter ",
+			"detail": " q | ESC: Exit Detail ",
 		},
 	}
